refactor: Move DTO classes to appropriate packages for better organization and more package cleaning
- remove not use file
