+++ src/main/java/com/mycodingtest/authorization/dto/UserInfoResponse.java
@@ -1,4 +1,4 @@
-package com.mycodingtest.authorization;
+package com.mycodingtest.authorization.dto;
 
 public record UserInfoResponse(
         String picture,
