+++ src/main/java/com/mycodingtest/judgmentresult/JudgmentResultMapper.java
@@ -1,5 +1,7 @@
 package com.mycodingtest.judgmentresult;
 
+import com.mycodingtest.judgmentresult.dto.JudgmentResultResponse;
+
 public class JudgmentResultMapper {
 
     public static JudgmentResultResponse toResponse(JudgmentResult entity) {
