+++ src/main/java/com/mycodingtest/judgmentresult/JudgmentResultService.java
@@ -1,5 +1,7 @@
 package com.mycodingtest.judgmentresult;
 
+import com.mycodingtest.judgmentresult.dto.JudgmentResultResponse;
+import com.mycodingtest.judgmentresult.dto.JudgmentResultSaveRequest;
 import com.mycodingtest.review.Review;
 import com.mycodingtest.solvedproblem.SolvedProblem;
 import com.mycodingtest.user.User;
