+++ src/main/java/com/mycodingtest/review/dto/ReviewRatingLevelsUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.mycodingtest.review;
+package com.mycodingtest.review.dto;
 
 public record ReviewRatingLevelsUpdateRequest(
         int difficultyLevel,
