+++ src/main/java/com/mycodingtest/review/dto/ReviewStatusUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.mycodingtest.review;
+package com.mycodingtest.review.dto;
 
 import jakarta.validation.constraints.NotNull;
