--- src/main/java/com/mycodingtest/solvedproblem/MemoResponse.java
@@ -1,6 +1,0 @@
-package com.mycodingtest.solvedproblem;
-
-public record MemoResponse(
-        String content
-) {
-}
