--- src/main/java/com/mycodingtest/solvedproblem/MemoSaveRequest.java
@@ -1,6 +1,0 @@
-package com.mycodingtest.solvedproblem;
-
-public record MemoSaveRequest(
-        String content
-) {
-}
