--- src/main/java/com/mycodingtest/solvedproblem/SolvedProblemResponse.java
@@ -1,10 +1,0 @@
-package com.mycodingtest.solvedproblem;
-
-import java.time.LocalDateTime;
-
-public record SolvedProblemResponse(
-        LocalDateTime recentSubmitAt,
-        boolean isReviewed,
-        boolean isFavorite
-) {
-}
