+++ src/main/java/com/mycodingtest/solvedproblem/SolvedProblemService.java
@@ -3,6 +3,8 @@ package com.mycodingtest.solvedproblem;
 import com.mycodingtest.common.exception.ResourceNotFoundException;
 import com.mycodingtest.judgmentresult.JudgmentResult;
 import com.mycodingtest.judgmentresult.JudgmentResultRepository;
+import com.mycodingtest.solvedproblem.dto.SolvedProblemWithReviewResponse;
+import com.mycodingtest.tag.MyTagListResponse;
 import com.mycodingtest.storage.StorageService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
