+++ src/main/java/com/mycodingtest/tag/MyTagListResponse.java
@@ -1,4 +1,4 @@
-package com.mycodingtest.solvedproblem;
+package com.mycodingtest.tag;
 
 public record MyTagListResponse(
         Object[] tagIds
