+++ src/main/java/com/mycodingtest/tag/dto/AlgorithmTagResponse.java
@@ -1,4 +1,4 @@
-package com.mycodingtest.solvedproblem;
+package com.mycodingtest.tag.dto;
 
 public record AlgorithmTagResponse(
         Object[] tagIds
